add fuzz for openvino about keras model

+++ keras/fuzz/run_fuzz.py
@@ -3,16 +3,10 @@ import sys
 import subprocess
 import multiprocessing
 import datetime
+import shutil
 
 import warnings
 warnings.filterwarnings('ignore')
-
-keras_run_head = """from test_tvm_keras import layer_test
-from tensorflow import keras
-import tensorflow as tf
-import numpy as np
-
-"""
 
 
 def run_subprocess(python_program):
@@ -26,7 +20,7 @@ def run_subprocess(python_program):
             output_final += line
         print(output_final)
         return
-    else:
+    else:  # invalid test cases
         err_output = run_flag.stderr.decode('utf-8')
         output_final = ''
         for line in err_output.split("\\n"):
@@ -37,31 +31,38 @@ def run_subprocess(python_program):
     return
 
 
-def gen_test_case_file(test_str, save_dir, test_id):
+def gen_test_case_file(SUT, test_str, save_dir, test_id):
+    keras_run_head = f"""from test_{SUT}_keras import layer_test
+from tensorflow import keras
+import tensorflow as tf
+import numpy as np
+
+"""
     test_str = keras_run_head + test_str
     if not os.path.exists(save_dir):
         os.makedirs(save_dir)
-    os.system(f"cp test_tvm_keras.py {save_dir}")
+    shutil.copy(f"test_{SUT}_keras.py", save_dir)
+
     save_test_file_path = os.path.join(save_dir, f'test_{test_id}.py')
-    with open(save_test_file_path, 'w', encoding='utf-8') as test_f:
-        test_f.write(test_str)
+    if not os.path.exists(save_test_file_path):
+        with open(save_test_file_path, 'w', encoding='utf-8') as test_f:
+            test_f.write(test_str)
     return save_test_file_path
 
 
-def run_all_test(test_file, begin_id=1):
-    if 'docter' in test_file:
-        begin_id = 20993
+def run_all_test(test_file, SUT, begin_id=1):
     with open(test_file, 'r', encoding='utf-8') as f:
         all_lines = f.readlines()
 
     all_test_cases = []
     for cnt, line in enumerate(all_lines):
+        tc_id = cnt+begin_id
         line = line.strip()
         if line.startswith("layer_test"):
-            line = line.strip()[:-1] + f"count={cnt+begin_id},)"  # give correct bug_id
-
-            save_test_file_path = gen_test_case_file(line, "all_all", cnt+begin_id)
+            tc_line = line.strip()[:-1] + f"count={tc_id},)"  # give correct bug_id
+            save_test_file_path = gen_test_case_file(SUT, tc_line, "all_keras_tc", tc_id)
             all_test_cases.append(save_test_file_path)
+            # execute the all test cases
             run_subprocess(save_test_file_path)
 
     #pool = multiprocessing.Pool(processes=1)#multiprocessing.cpu_count()//2)
@@ -73,10 +74,8 @@ def run_all_test(test_file, begin_id=1):
 
 if __name__ == '__main__':
     starttime = datetime.datetime.now()
-    # collected_test_cases_file = f"borrow_all_test.py"
-    # collected_test_cases_file = f"borrow_all_test_deduplicate.py"
-    collected_test_cases_file = sys.argv[1]
-    run_all_test(collected_test_cases_file)
-
+    collected_test_cases_file = sys.argv[1]   # ../data/combined_sources_keras_test_41986.py
+    SUT = sys.argv[2]  # [tvm, openvino]
+    run_all_test(collected_test_cases_file, SUT)
     endtime = datetime.datetime.now()
     print("Finish all, time consuming(min): ", (endtime - starttime).seconds/60)
